重构部分代码

+++ hz-back/src/main/java/com/beeasy/hzback/modules/setting/dao/IDepartmentDao.java
@@ -1,14 +1,11 @@
 package com.beeasy.hzback.modules.setting.dao;
 
 import com.beeasy.hzback.modules.setting.entity.Department;
-import com.beeasy.hzback.modules.setting.entity.User;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Set;
 
-public interface IDepartmentDao extends JpaRepository<Department,Integer> {
+public interface IDepartmentDao extends JpaRepository<Department,Long> {
     List<Department> findAllByParent(Department department);
     List<Department> findAllByParentId(Integer parentId);
     Department findByParent(Department department);
