+++ hz-back/src/main/java/com/beeasy/hzback/modules/system/form/QuartersEdit.java
@@ -3,15 +3,12 @@ package com.beeasy.hzback.modules.system.form;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
-import org.hibernate.validator.constraints.NotEmpty;
-
-import javax.validation.constraints.NotNull;
 
 @ApiModel
 @Data
 public class QuartersEdit {
     @ApiModelProperty(value = "岗位ID", required = true)
-    Integer id;
+    Long id;
 
     @ApiModelProperty(value = "岗位名称")
     private String name;
